+++ devicemanagement/device_compliance_policies_item_scheduled_actions_for_rule_request_builder.go
@@ -14,7 +14,7 @@ import (
 type DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilder struct {
     i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.BaseRequestBuilder
 }
-// DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilderGetQueryParameters the list of scheduled action per rule for this compliance policy. This is a required property when creating any individual per-platform compliance policies.
+// DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilderGetQueryParameters the list of scheduled action for this rule
 type DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilderGetQueryParameters struct {
     // Include count of items
     Count *bool `uriparametername:"%24count"`
@@ -29,9 +29,9 @@ type DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilderGetQueryPa
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Skip the first n items
-    Skip *int32 `uriparametername:"%24skip"`
+    Skip *int64 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int32 `uriparametername:"%24top"`
+    Top *int64 `uriparametername:"%24top"`
 }
 // DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilderGetRequestConfiguration struct {
@@ -79,7 +79,7 @@ func NewDeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilder(rawUrl
 func (m *DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilder) Count()(*DeviceCompliancePoliciesItemScheduledActionsForRuleCountRequestBuilder) {
     return NewDeviceCompliancePoliciesItemScheduledActionsForRuleCountRequestBuilderInternal(m.BaseRequestBuilder.PathParameters, m.BaseRequestBuilder.RequestAdapter)
 }
-// Get the list of scheduled action per rule for this compliance policy. This is a required property when creating any individual per-platform compliance policies.
+// Get the list of scheduled action for this rule
 // returns a DeviceComplianceScheduledActionForRuleCollectionResponseable when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
 func (m *DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilder) Get(ctx context.Context, requestConfiguration *DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilderGetRequestConfiguration)(ie233ee762e29b4ba6970aa2a2efce4b7fde11697ca9ea81099d0f8269309c1be.DeviceComplianceScheduledActionForRuleCollectionResponseable, error) {
@@ -119,7 +119,7 @@ func (m *DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilder) Post
     }
     return res.(ie233ee762e29b4ba6970aa2a2efce4b7fde11697ca9ea81099d0f8269309c1be.DeviceComplianceScheduledActionForRuleable), nil
 }
-// ToGetRequestInformation the list of scheduled action per rule for this compliance policy. This is a required property when creating any individual per-platform compliance policies.
+// ToGetRequestInformation the list of scheduled action for this rule
 // returns a *RequestInformation when successful
 func (m *DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilder) ToGetRequestInformation(ctx context.Context, requestConfiguration *DeviceCompliancePoliciesItemScheduledActionsForRuleRequestBuilderGetRequestConfiguration)(*i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.RequestInformation, error) {
     requestInfo := i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.NewRequestInformationWithMethodAndUrlTemplateAndPathParameters(i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.GET, m.BaseRequestBuilder.UrlTemplate, m.BaseRequestBuilder.PathParameters)
