+++ directory/feature_rollout_policies_item_applies_to_count_request_builder.go
@@ -44,9 +44,9 @@ func NewFeatureRolloutPoliciesItemAppliesToCountRequestBuilder(rawUrl string, re
 }
 // Get get the number of the resource
 // Deprecated: Feature Rollout Policies have been grouped with other policies under /policies. The existing /directory/featureRolloutPolicies is deprecated and will stop returning data on 06/30/2021. Please use /policies/featureRolloutPolicies. as of 2021-01/DirectoryFeatureRolloutPolicies on 2021-03-05 and will be removed 2021-06-30
-// returns a *int32 when successful
+// returns a *int64 when successful
 // returns a ODataError error when the service returns a 4XX or 5XX status code
-func (m *FeatureRolloutPoliciesItemAppliesToCountRequestBuilder) Get(ctx context.Context, requestConfiguration *FeatureRolloutPoliciesItemAppliesToCountRequestBuilderGetRequestConfiguration)(*int32, error) {
+func (m *FeatureRolloutPoliciesItemAppliesToCountRequestBuilder) Get(ctx context.Context, requestConfiguration *FeatureRolloutPoliciesItemAppliesToCountRequestBuilderGetRequestConfiguration)(*int64, error) {
     requestInfo, err := m.ToGetRequestInformation(ctx, requestConfiguration);
     if err != nil {
         return nil, err
@@ -54,14 +54,14 @@ func (m *FeatureRolloutPoliciesItemAppliesToCountRequestBuilder) Get(ctx context
     errorMapping := i2ae4187f7daee263371cb1c977df639813ab50ffa529013b7437480d1ec0158f.ErrorMappings {
         "XXX": i20a3050780ee0b0cde0a884a4f35429a20d60067e3bcda382ec5400079147459.CreateODataErrorFromDiscriminatorValue,
     }
-    res, err := m.BaseRequestBuilder.RequestAdapter.SendPrimitive(ctx, requestInfo, "int32", errorMapping)
+    res, err := m.BaseRequestBuilder.RequestAdapter.SendPrimitive(ctx, requestInfo, "int64", errorMapping)
     if err != nil {
         return nil, err
     }
     if res == nil {
         return nil, nil
     }
-    return res.(*int32), nil
+    return res.(*int64), nil
 }
 // ToGetRequestInformation get the number of the resource
 // Deprecated: Feature Rollout Policies have been grouped with other policies under /policies. The existing /directory/featureRolloutPolicies is deprecated and will stop returning data on 06/30/2021. Please use /policies/featureRolloutPolicies. as of 2021-01/DirectoryFeatureRolloutPolicies on 2021-03-05 and will be removed 2021-06-30
