+++ financials/companies_item_sales_quotes_item_sales_quote_lines_item_item_escaped_picture_request_builder.go
@@ -30,9 +30,9 @@ type CompaniesItemSalesQuotesItemSalesQuoteLinesItemItem_EscapedPictureRequestBu
     // Select properties to be returned
     Select []string `uriparametername:"%24select"`
     // Skip the first n items
-    Skip *int32 `uriparametername:"%24skip"`
+    Skip *int64 `uriparametername:"%24skip"`
     // Show only the first n items
-    Top *int32 `uriparametername:"%24top"`
+    Top *int64 `uriparametername:"%24top"`
 }
 // CompaniesItemSalesQuotesItemSalesQuoteLinesItemItem_EscapedPictureRequestBuilderGetRequestConfiguration configuration for the request such as headers, query parameters, and middleware options.
 type CompaniesItemSalesQuotesItemSalesQuoteLinesItemItem_EscapedPictureRequestBuilderGetRequestConfiguration struct {
